+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Refactor defaults to attributes, it's more self documented now
 - Move chef version to configuration files, fixing rails5 chef problem
 - templates displayed at help generated with yml information
 - refactor templates file structure: one yml per template